debugging infinite loops not completed

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: rhernand <rhernand@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/17 18:40:29 by rhernand          #+#    #+#             */
-/*   Updated: 2024/06/19 12:50:13 by rhernand         ###   ########.fr       */
+/*   Updated: 2024/06/21 13:03:08 by rhernand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t count, size_t size);
 size_t	ft_strlen(const char *s);
 char	*ft_strjoin(const char *s1, const char *s2);
-size_t	ft_strchr(const char *s, int c);
+int		ft_strchr(const char *s, int c);
 size_t	ft_strlcpy(char *dest, char *src, size_t dstsize);
 
 #endif
